put runner into separate file
